🚚 move queries accordingly to their scope
